feat(build): v3 slice 4-5 — CI workflow (AC-16) + operator runbooks (AC-21)
Slice 4 (AC-16): GitHub Actions CI (.github/workflows/ci.yml) with fmt, clippy
-D warnings, test against Postgres 16 service container, and cargo-deny
(advisories/bans/licenses/sources) jobs. deny.toml with license allowlist
(Apache-2.0, MIT, BSD-*, ISC, MPL-2.0, OpenSSL, Unicode-3.0) and
unknown-registry/git = "deny". Resolved 9 clippy lints: redundant closures in
api/messages.rs + models/llm_call.rs, unnecessary i32/i64 casts in
background/stale.rs + runtime/wake_loop.rs, #[allow(clippy::too_many_arguments)]
on append_event + insert_llm_call, assert_eq!(.., bool) -> assert! in
agent_mgmt_test. cargo clippy --all-targets -- -D warnings: exit 0.
cargo fmt --all -- --check: exit 0.

Slice 5 (AC-21): 5 operator runbooks under docs/runbooks/ —
stale-wake-triage, db-restore, migration-rollback, rate-limit-tuning,
webhook-debugging — each with Symptom / Diagnostic Commands (concrete
psql/curl/docker compose/openssl invocations) / Remediation / Escalation.

Verification: cargo test --all-targets -- --test-threads=1 against Postgres 16
on port 5433 -> 30 passed, 0 failed, EXIT=0.

+++ src/background/stale.rs
@@ -31,7 +31,7 @@ pub async fn start_stale_recovery(
             _ = interval.tick() => {}
         }
 
-        match agent::find_stale_agents(&pool, config.stale_wake_hours as i64).await {
+        match agent::find_stale_agents(&pool, config.stale_wake_hours).await {
             Ok(stale_agents) => {
                 for a in stale_agents {
                     info!(agent_id = %a.id, wake_id = ?a.wake_id, "Recovering stale agent");
@@ -40,10 +40,18 @@ pub async fn start_stale_recovery(
                     } else {
                         // Record stale_wake_recovery event
                         let _ = event::append_event(
-                            &pool, a.id, "stale_wake_recovery", "system",
-                            a.wake_id, None, None, None,
-                            Some("Agent recovered from stale wake"), None,
-                        ).await;
+                            &pool,
+                            a.id,
+                            "stale_wake_recovery",
+                            "system",
+                            a.wake_id,
+                            None,
+                            None,
+                            None,
+                            Some("Agent recovered from stale wake"),
+                            None,
+                        )
+                        .await;
                     }
                 }
             }
